+++ command_context_windows.go
@@ -1,0 +1,10 @@
+//go:build windows
+
+package oneagent
+
+import (
+	"context"
+	"os/exec"
+)
+
+func applyCancelableCommandContext(_ *exec.Cmd, _ context.Context) {}